Refactor player hand components and fix type/lint errors
- Split PlayerHandAnimated into separate HumanHandAnimated and AIHandAnimated components
- Extract redundant styles into shared styles file
- Fix TypeScript errors in test files (proper imports and type annotations)
- Fix ESLint warnings for unused imports and React Hook dependencies
- Remove unused GameScreen wrapper component
- Fix Android StatusBar styling and positioning
- Update app entry point to use GameScreenController directly
- Update architecture documentation to reflect component changes
- Improve code organization and separation of concerns
